+++ Sources/AXorcist/Utils/String+HelperExtensions.swift
@@ -1,11 +1,6 @@
 import Foundation
 
-/// String extension from Scanner
 extension String {
-    subscript(offset: Int) -> Character {
-        self[index(startIndex, offsetBy: offset)]
-    }
-
     func truncated(to length: Int, trailing: String = "...") -> String {
         if self.count > length {
             String(self.prefix(length - trailing.count)) + trailing